chore: clear clippy warning backlog and enforce -D warnings in CI
Allow collapsible_if, too_many_arguments, and collapsible_match at the
workspace level; collapsible_match's auto-fix moves conditions into
match guards, which changes key-dispatch fall-through, so it must stay
off. Fix the remaining warnings by hand.

Includes a real fix in number_ops: octal literal scanning accepted the
digits 8 and 9, which made from_str_radix(_, 8) fail and Ctrl-A treat
values like 0o19 as 0.

Also drop a needless Renderer::new() from emoji_tab_rendering_test —
it requires a real TTY and fails in CI.

+++ ovim-core/src/editor/decoration.rs
@@ -639,7 +639,7 @@ where
         .iter()
         .map(|hint| {
             let line = hint.position.line as usize;
-            let utf16_col = hint.position.character as u32;
+            let utf16_col = hint.position.character;
             // Convert UTF-16 offset → char index.
             let text_for_line = line_text(line);
             let char_idx = crate::lsp::utf16_to_char_col(&text_for_line, utf16_col);
